Implementation of function calls and return values.
A map remembers the callbacks for the corresponding sequence
of the function call.

VimSocketInterfaceCommunicator fixed parsing when the reply
consists solely of the sequence number.

VimFunctions parse their argument themselves (with default implementation).
This permits a function to accept varying result types and count.

VimValue can be constructed empty, with INVALID type.
It is modified through the default operator =.