Fix -cc1 detection, and argc/argv usage: now preprocesses!
I'll have to remember one day, that argc is the number of elements of
argv, *including* the executable name argv[0] in the counting...

With this confusion, I was checking the executable against "-cc1",
which was giving the arguments to the wrong handler function.
Moreover, I was not giving the very last argument to the
CompilerInvocation, hence it was waiting for stdin data...

Print the gotten arguments.