Fix missing <stddef.h>
input/input03.c includes <stdio.h>, which in turns includes <stddef.h>.
This file is to be shipped with the compiler, clang uses the a path
relative to the executable by default, hence it takes
/usr/{local/}bin/[clang], and appends ../lib/clang/3.1.
However our programs are not located aside the header distribution,
so I hardcoded /usr/local/lib/clang/3.1 to fix the problem.

Note that this assumes you installed the compiled LLVM, and didn't use
--path /usr in ./configure.
Another solution would be to leave symlinks in
test/amnoid-clang-tutorial/{,dist/}, and to perform some smart
resolution in the code.